Many structural changes to allow customization
  * Moved the functions computing the number of points / lines to synchronize on
during the current iteration in increase.cpp / increase.h. As such, we can
extract the functions and their parameters when running a test
  * Moved the dimensions of the problem and the related configuration value to
dynamic_defines.h. This file is generated by the generate_dynamic_defines.py
script (will output default values if none is provided). As such, we can now
change the values of the problem as we want at the cost of one build
  * Also, why did I use CMake instead of a plain old Makefile ? How am I
supposed to add "If this file does not exist when running make then call the
Python script" ?