We now have performance
  * Forcibly push early from Deduplicate to both Compress and Reorder if
we have too many items to insert into one of the FIFO. Basically, if we
insert more than step items in one FIFO then we forcibly push into the
other to make sure there is no stagnation because one of the FIFOs does
not have enough work to actually mandate a push.
  * Reorder no longer writes on the disk as it receives data, instead it
puts everything inside the tree as it receives data, and only once it
has received everything does it actually write on the disk. This removes
most of the incomprehensible wait times we had before.
  * The first change applies only to the SmartFIFO version of the code
because it would not make sense to add this optimisation to the original
version (this optimisation is, after all, specific to SmartFIFO)
  * The second optimisation is applied to both versions as it is a major
deviation from the original code that does not exclusively depends on a
new programming abstraction
  * Added lots of instrumentation to the code to make it easy to have
log and study what is happening. Most of these instrumentations are
disabled for now as we are focused on evaluting performance in real
conditions.