Some minor fixes
  * The Reorder input observers now properly receive the correct amount
of data. This was due to the old code not properly counting the amount
of data inserted in the observer by FIFO, but instead counting the
amount of data inserted between all FIFOs
  * Added some logs to make sure that every layer receives the correct
amount of data (zdiffed the outputs to make sure everything is okay)
  * Added OBS_LIMIT in encode_naive_queue.cpp that is used to set
limits. This can be used to completely disable reconfigurations